+++ app/build.gradle
@@ -41,14 +41,14 @@ repositories {
 
 dependencies {
     implementation project(path: ':library')
-    implementation project(path: ':library-extension-rx')
-//    implementation 'com.github.instacart.truetime-android:library-extension-rx:master-SNAPSHOT'
+//    implementation project(path: ':library-extension-rx')
+    implementation 'com.github.instacart.truetime-android:library-extension-rx:3.5'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
 
-    implementation 'androidx.appcompat:appcompat:1.0.2'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.5.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
     implementation "io.reactivex.rxjava2:rxandroid:2.1.1"
 }
