--- library-extension-rx/build.gradle
@@ -1,37 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-android {
-    compileSdkVersion rootProject.ext.compileSdkVersion
-
-    defaultConfig {
-        minSdkVersion rootProject.ext.minSdkVersion
-
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    buildFeatures {
-        buildConfig = false
-    }
-}
-
-
-repositories {
-  mavenCentral()
-}
-
-dependencies {
-    api "org.jetbrains.kotlin:kotlin-stdlib"
-    api 'io.reactivex.rxjava2:rxjava:2.2.8'
-    api project(path: ':library')
-}
-
-group='com.github.instacart'
