--- library/src/main/java/com/instacart/truetime/cache/TrueTimeCacheProvider.kt
@@ -1,11 +1,0 @@
-package com.instacart.truetime.cache
-
-interface TrueTimeCacheProvider {
-    fun clear()
-
-    fun update(
-        bootTime: Long? = null,
-        deviceUptime: Long? = null,
-        sntpTime: Long? = null
-    )
-}
