+++ library/src/test/java/com/instacart/truetime/time/TrueTimeImplTest.kt
@@ -1,0 +1,88 @@
+package com.instacart.truetime.time
+
+class TrueTimeImplTest {
+
+  // ntp single ip : 216.239.35.4
+  /**
+   *
+
+  0 = 1606352796468
+  1 = 1606352795894
+  2 = 1606352795894
+  3 = 1606352796546
+  4 = 0
+  5 = 10
+  6 = 1
+  7 = 704540455
+
+
+  0 = 1606352813766
+  1 = 1606352813177
+  2 = 1606352813177
+  3 = 1606352813805
+  4 = 0
+  5 = 6
+  6 = 1
+  7 = 704557715
+
+  0 = 1606352816644
+  1 = 1606352816116
+  2 = 1606352816116
+  3 = 1606352816832
+  4 = 0
+  5 = 6
+  6 = 1
+  7 = 704560742
+
+  0 = 1606352818690
+  1 = 1606352818116
+  2 = 1606352818116
+  3 = 1606352818761
+  4 = 0
+  5 = 11
+  6 = 1
+  7 = 704562670
+
+  sorted by clock offset
+
+  0 = 1606352816644
+  1 = 1606352816116
+  2 = 1606352816116
+  3 = 1606352816832
+  4 = 0
+  5 = 6
+  6 = 1
+  7 = 704560742
+
+  0 = 1606352796468
+  1 = 1606352795894
+  2 = 1606352795894
+  3 = 1606352796546
+  4 = 0
+  5 = 10
+  6 = 1
+  7 = 704540455
+
+
+  0 = 1606352818690
+  1 = 1606352818116
+  2 = 1606352818116
+  3 = 1606352818761
+  4 = 0
+  5 = 11
+  6 = 1
+  7 = 704562670
+
+  0 = 1606352813766
+  1 = 1606352813177
+  2 = 1606352813177
+  3 = 1606352813805
+  4 = 0
+  5 = 6
+  6 = 1
+  7 = 704557715
+
+
+
+   */
+}
